Restructure crate and make it compile on no std
