+++ crossbeam-channel/Cargo.toml
@@ -15,6 +15,13 @@ description = "Multi-producer multi-consumer channels for message passing"
 keywords = ["channel", "mpmc", "select", "golang", "message"]
 categories = ["algorithms", "concurrency", "data-structures"]
 
+[features]
+no_std = []
+nightly = []
+
+[dependencies]
+cfg-if = "0.1.2"
+
 [dependencies.crossbeam-utils]
 version = "0.6.5"
 path = "../crossbeam-utils"
