+++ crossbeam-channel/src/lib.rs
@@ -347,32 +347,41 @@
 #![warn(missing_docs)]
 #![warn(missing_debug_implementations)]
 
+#![cfg_attr(feature = "no_std", no_std)]
+#![cfg_attr(feature = "nightly", feature(const_fn))]
+
+#[macro_use]
+extern crate cfg_if;
 extern crate crossbeam_utils;
 
-mod channel;
-mod context;
-mod counter;
-mod err;
-mod flavors;
-mod select;
-mod select_macro;
-mod utils;
-mod waker;
+cfg_if! {
+    if #[cfg(not(feature = "no_std"))] {
+        mod channel;
+        mod context;
+        mod counter;
+        mod err;
+        mod flavors;
+        mod select;
+        mod select_macro;
+        mod utils;
+        mod waker;
 
-/// Crate internals used by the `select!` macro.
-#[doc(hidden)]
-pub mod internal {
-    pub use select::SelectHandle;
-    pub use select::{select, select_timeout, try_select};
+        /// Crate internals used by the `select!` macro.
+        #[doc(hidden)]
+        pub mod internal {
+            pub use select::SelectHandle;
+            pub use select::{select, select_timeout, try_select};
+        }
+
+        pub use channel::{after, never, tick};
+        pub use channel::{bounded, unbounded};
+        pub use channel::{IntoIter, Iter, TryIter};
+        pub use channel::{Receiver, Sender};
+
+        pub use select::{Select, SelectedOperation};
+
+        pub use err::{ReadyTimeoutError, SelectTimeoutError, TryReadyError, TrySelectError};
+        pub use err::{RecvError, RecvTimeoutError, TryRecvError};
+        pub use err::{SendError, SendTimeoutError, TrySendError};
+    }
 }
-
-pub use channel::{after, never, tick};
-pub use channel::{bounded, unbounded};
-pub use channel::{IntoIter, Iter, TryIter};
-pub use channel::{Receiver, Sender};
-
-pub use select::{Select, SelectedOperation};
-
-pub use err::{ReadyTimeoutError, SelectTimeoutError, TryReadyError, TrySelectError};
-pub use err::{RecvError, RecvTimeoutError, TryRecvError};
-pub use err::{SendError, SendTimeoutError, TrySendError};
