+++ crossbeam-epoch/Cargo.toml
@@ -16,14 +16,15 @@ keywords = ["lock-free", "rcu", "atomic", "garbage"]
 categories = ["concurrency", "memory-management", "no-std"]
 
 [features]
-default = ["std"]
+default = []
 nightly = ["crossbeam-utils/nightly", "arrayvec/use_union"]
-std = ["crossbeam-utils/std", "lazy_static"]
+no_std = ["crossbeam-utils/no_std", "lazy_static/spin_no_std"]
 sanitize = [] # Makes it more likely to trigger any potential data races.
 
 [dependencies]
 cfg-if = "0.1.2"
 memoffset = "0.2"
+lazy_static = { version = "^1" }
 
 [dependencies.arrayvec]
 version = "0.4"
@@ -33,10 +34,6 @@ default-features = false
 version = "0.6"
 path = "../crossbeam-utils"
 default-features = false
-
-[dependencies.lazy_static]
-version = "1"
-optional = true
 
 [dependencies.scopeguard]
 version = "0.3"
