+++ crossbeam-queue/src/array_queue.rs
@@ -14,6 +14,7 @@ use core::marker::PhantomData;
 use core::mem;
 use core::ptr;
 use core::sync::atomic::{self, AtomicUsize, Ordering};
+use alloc::vec::Vec;
 
 use crossbeam_utils::{Backoff, CachePadded};
