+++ crossbeam-queue/src/seg_queue.rs
@@ -4,6 +4,7 @@ use core::marker::PhantomData;
 use core::mem::{self, ManuallyDrop};
 use core::ptr;
 use core::sync::atomic::{self, AtomicPtr, AtomicUsize, Ordering};
+use alloc::boxed::Box;
 
 use crossbeam_utils::{Backoff, CachePadded};
