+++ crossbeam-skiplist/src/lib.rs
@@ -2,15 +2,15 @@
 
 #![warn(missing_docs)]
 #![warn(missing_debug_implementations)]
-#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(feature = "no_std", no_std)]
 
 #[macro_use]
 extern crate cfg_if;
-#[cfg(feature = "std")]
+#[cfg(not(feature = "no_std"))]
 extern crate core;
 
 cfg_if! {
-    if #[cfg(feature = "nightly")] {
+    if #[cfg(feature = "no_std")] {
         extern crate alloc;
     } else {
         mod alloc {
@@ -29,7 +29,7 @@ pub mod base;
 pub use base::SkipList;
 
 cfg_if! {
-    if #[cfg(feature = "std")] {
+    if #[cfg(not(feature = "no_std"))] {
         pub mod map;
         #[doc(inline)]
         pub use map::SkipMap;
