+++ crossbeam-utils/Cargo.toml
@@ -16,13 +16,13 @@ keywords = ["scoped", "thread", "atomic", "cache"]
 categories = ["algorithms", "concurrency", "data-structures", "no-std"]
 
 [features]
-default = ["std"]
+default = []
 nightly = []
-std = ["lazy_static"]
+no_std = ["lazy_static/spin_no_std"]
 
 [dependencies]
 cfg-if = "0.1.2"
-lazy_static = { version = "1.1.0", optional = true }
+lazy_static = { version = "^1" }
 
 [dev-dependencies]
 rand = "0.6"
