+++ crossbeam-utils/src/lib.rs
@@ -28,16 +28,18 @@
 
 #![warn(missing_docs)]
 #![warn(missing_debug_implementations)]
-#![cfg_attr(not(feature = "std"), no_std)]
+#![cfg_attr(feature = "no_std", no_std)]
 #![cfg_attr(feature = "nightly", feature(cfg_target_has_atomic))]
 
 #[macro_use]
 extern crate cfg_if;
-#[cfg(feature = "std")]
+#[cfg(not(feature = "no_std"))]
 extern crate core;
+#[macro_use]
+extern crate lazy_static;
 
 cfg_if! {
-    if #[cfg(feature = "nightly")] {
+    if #[cfg(feature = "no_std")] {
         extern crate alloc;
     } else {
         mod alloc {
@@ -56,10 +58,7 @@ mod backoff;
 pub use backoff::Backoff;
 
 cfg_if! {
-    if #[cfg(feature = "std")] {
-        #[macro_use]
-        extern crate lazy_static;
-
+    if #[cfg(not(feature = "no_std"))] {
         pub mod sync;
         pub mod thread;
     }
